We introduce IComputeNode<T>.
IComputeScalar now inherits from IComputeNode<float>
All consuming locations where IComputeScalar was expected, now only expect a IComputeNode<float>
IComputeColor now inherits from IComputeNode<Vector4>
All consuming locations where IComputeColor was expected, now only expect a IComputeNode<Vector4>

The HasChanged property moved up from IComputeColor to IComputeNode, got implemented on more places, but no code was changed to make use of the new feature that every IComputeNode now can be asked for whether it changed.

All classes that implemented IComputeScalar or IComputeColor still do so.
By that no user code will break as the richer interface stays implemented.
However it is debatable if these interfaces are still necessary in the longterm.

In the long run we also might be able to remove ComputeBinaryColor and ComputeBinaryScalar, ComputeShaderClassScalar and ComputeShaderClassColor.

By postponing these removals this is a quite conservative change.
However we changed two bits that might break user code. For the case that someone made use of ComputeBinaryBase<T> or ComputeShaderClassBase<T>. T now is not about the node, but the data flowing through the node.

Debatable: should we use IComputeNode<Vector4> or IComputeNode<Color4> when it's about colors or normal maps? I went for Vector4, since it's more general, but it might be nice to distinguish them and have operators in the compute node graph that convert between them (even though this would not result in any shader instruction getting added to the shader code). By that you have some more type awareness in your compute node graph. It might however well be that it just gets in the way all the time and so IComputeNode<Vector4> is the proposed standard way to deal with shaders that return float4.